+++ apis/vara-eth/CHANGELOG.md
@@ -5,3 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
+
+### Changed
+
+- Expose contract addresses from mirror/router/wvara clients in https://github.com/gear-tech/gear-js/pull/2197
+
+### Added
+
+- Pick the next validator as a recipient for injected transactions in https://github.com/gear-tech/gear-js/pull/2205